refactor: 공통된 경로 구하기 로직 부모 클래스로 추상화
