+++ src/main/java/model/piece/Pho.java
@@ -2,7 +2,6 @@ package model.piece;
 
 import java.util.Map;
 import java.util.Map.Entry;
-import model.Path;
 import model.Point;
 import model.Team;
 
@@ -12,29 +11,11 @@ public class Pho extends Piece {
         super(team,PieceName.PHO);
     }
 
-    public boolean isValidPoint(Point beforePoint, Point targetPoint) {
-        boolean isStraightMove = beforePoint.x() == targetPoint.x()
-                || beforePoint.y() == targetPoint.y();
-        boolean isSamePoint = beforePoint.x() == targetPoint.x()
-                && beforePoint.y() == targetPoint.y();
-        return isStraightMove && !isSamePoint;
-    }
-
     @Override
-    public Path calculatePath(Point beforePoint, Point targetPoint) {
+    public boolean isValidPoint(Point beforePoint, Point targetPoint) {
         int vectorX = getVectorX(beforePoint, targetPoint);
         int vectorY = getVectorY(beforePoint, targetPoint);
-
-        int unitVectorX = getUnitVector(vectorX);
-        int unitVectorY = getUnitVector(vectorY);
-
-        Path path = new Path();
-
-        for (int i = 0; i < Math.max(vectorY, vectorX); i++) {
-            path.addPoint(new Point(targetPoint.x() - unitVectorX * i, targetPoint.y() - unitVectorY * i));
-        }
-
-        return path;
+        return (vectorX == 0) ^ (vectorY == 0);
     }
 
     @Override
